backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
apscheduler
httpx
python-multipart
bcrypt==4.0.1
pytz